extract api logic into its own directory
