+++ frontend/src/features/users/index.ts
@@ -1,4 +1,3 @@
-export * from "./api/userApi";
 export * from "./components/EditUserProfileForm";
 export * from "./components/OnboardingForm";
 export * from "./types/user.type";
